client: add 'hidden' property to show hidden changesets.
This enables interactions with the obsolete changesets in the repository:

- add the attribute in client class
- add the keyword to the relevant commands
- enable log without hidden changesets even when self.hidden is True
- add a few tests with the hidden keyword

This changeset mirrors the behavior of the mercurial global command --hidden:
an attribute is added to the client library. If set at True, adds the hidden
keyword to all command which can use it to show hidden changesets.

The alternative would be to add the keyword in rawcommand, but the hidden flag
is not relevant for commands such as add or branch.